optimize .install.sh

+++ .install.sh
@@ -26,15 +26,16 @@ echo "done"
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks from the homedir to any files in the ~/dotfiles directory specified in $files
 ################################
 for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/."$file" ~/dotfiles_old/
+    if [[ -f ~/."$file" ]]; then
+        echo "Moving any existing dotfiles from ~ to $olddir"
+        mv ~/."$file" ~/dotfiles_old/
+    fi
     echo "Creating symlink to $file in home directory."
     ln -sf $dir/"$file" ~/."$file"
 done
-
+# delete the REDEME.md symlink
 if [[ -f ~/.README.md ]]; then
-
-    echo "Delete .README.md"
+    echo "Delete .README.md symlink"
     rm ~/.README.md
 fi
